functional tool node
The node now works as expected, and does so in the editor, and can be refreshed at will.
Currently runs on hardcoded heightdata. Work remains to import code for the grid object.
A debug texture has been imported to replace the vertex colors.

Hexagon corner is now calculated on runtime, with apropriate offset for the angle.

The old testcode has been disabled. New testcode is required for the new content.